refactor(scripts): extract Metro dev-server and Android dev-client launch helpers
Move the Metro lifecycle + reach-home readiness helpers out of
scripts/mobile-screenshots.ts into scripts/lib/metro-dev-server.ts, and the
Android dev-client launch sequence (install, Metro, adb reverse, launch-to-home
with retries) out of scripts/mobile-android-shots.ts into
scripts/lib/android-dev-client.ts, so CI's screenshot capture can reuse the
same sequence without the lib layer importing an orchestrator script.

Android now also reverses the readiness port (the app's reach-home ping never
left the emulator before) and retries the dev-client launch up to 3 times.
