+++ scripts/__tests__/mobile-screenshots.test.ts
@@ -18,18 +18,16 @@ import {
   summariseScreenshotRender,
   writeCapturedScreenshots,
   isIpadScreenshotDevice,
-  metroDevClientUrl,
   parseArgs,
   renderMaestroFlowForIosDevice,
   resolveAppStoreLocaleTargets,
   resolveIosScreenshotDevices,
   rotationDegreesForIosOrientation,
-  SCREENSHOT_READY_PORT,
-  screenshotReadinessCount,
   validateIosAppLauncherUrl,
   type IosScreenshotDevice,
   type ScreenshotOptions,
 } from '../mobile-screenshots';
+import { metroDevClientUrl, SCREENSHOT_READY_PORT, screenshotReadinessCount } from '../lib/metro-dev-server';
 
 const phoneDevices = ['iPhone 16 Pro Max'];
 const ipadDevices = ['iPad Pro 13-inch (M5)', 'iPad Pro 11-inch (M5)'];
