+++ scripts/tsconfig.json
@@ -40,6 +40,8 @@
     "lib/ios-simulator-lease.ts",
     "lib/ios-build-export.ts",
     "lib/metro-host.ts",
+    "lib/metro-dev-server.ts",
+    "lib/android-dev-client.ts",
     "mobile-ios-run.ts",
     "mobile-build-sim-app.ts",
     "mobile-dev-start.ts",
